requirements.txt
pytest
pytest-md
pytest-emoji
pytest-asyncio
protobuf
pyzmq
numpy
urllib3 == 1.26.15
twine
wheel
bitshuffle
